backend/requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
bcrypt==3.2.2
billiard==4.2.1
boto3==1.28.0
botocore==1.31.85
Brotli==1.1.0
cachetools==5.5.2
celery==5.3.6
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.3
cssselect2==0.8.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.100.0
fonttools==4.58.1
google-auth==2.40.3
google-genai==1.19.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jmespath==1.0.1
kombu==5.5.4
langdetect==1.0.9
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pdf2image==1.16.3
Pillow==10.0.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.6.1
pydantic_core==2.33.2
pydyf==0.11.0
pyphen==0.17.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==4.5.5
requests==2.32.3
rsa==4.9.1
s3transfer==0.6.2
sentencepiece==0.1.99
resend
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.24
starlette==0.27.0
tinycss2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.0.7
uvicorn==0.23.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
weasyprint==58.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
pypdf~=4.2
python-docx~=1.1
chardet~=5.2
python-magic~=0.4.27
Jinja2~=3.1